Fix LossySetitemError: cast margin_home/away to float before shootout assignment
Newer pandas raises instead of silently upcasting an int64 column when
assigning the float shootout_margin (0.5). zidane/cobi already cast to
float here; messi.py had drifted from that shared pattern.

+++ messi.py
@@ -919,8 +919,8 @@ shootout_mask = (
     (df['raw_margin_home'] == 0)
 )
 
-df['margin_home'] = df['raw_margin_home']
-df['margin_away'] = df['raw_margin_away']
+df['margin_home'] = df['raw_margin_home'].astype(float)
+df['margin_away'] = df['raw_margin_away'].astype(float)
 
 df.loc[shootout_mask & (df['shootout_winner'] == df['home_team']), 'margin_home'] =  shootout_margin
 df.loc[shootout_mask & (df['shootout_winner'] == df['home_team']), 'margin_away'] = -shootout_margin
